separate unit test for small mol & updated license
